Move cli/util flags based stuff into a cli/flags package.
